+++ src/Ouzo/Core/Db/Dialect/Sqlite3Dialect.php
@@ -100,4 +100,9 @@ class Sqlite3Dialect extends Dialect
     {
         throw new InvalidArgumentException("DISTINCT ON is not supported in sqlite3");
     }
+
+    protected function fromForDistinctCount(): string
+    {
+        throw new InvalidArgumentException("DISTINCT for COUNT is not supported in sqlite3");
+    }
 }